api/requirements.txt
fastapi
uvicorn
google-generativeai
groq